Added support for embedded documents and DbRefs

+++ src/main/java/com/mysmartfridge/domain/products/Product.java
@@ -20,7 +20,7 @@ public class Product {
 	 * tid of the product.
 	 */
 	@Id
-	private Long tid;
+	private Integer tid;
 	
 	/**
 	 * Name of the product/
